Removal of deprecated core modules / themes
THere are the usual core modules and themes enabled in 9.x that we can't use in 10.x so let's disable.
I've moved us to Claro for the admin theme and totally removed ckeditor from the content editing text filter. We don't have a lot of content so editing that without ckeditor help isn't a burden.