Upgrade to Drupal 10.2 (#322)
* Upgrade to Drupal 10.2

Upgrading the release to 10 means we will need this to work in two scenarios; that of a clean install AND that of an already running site upgrading in situ. We will need to test both scenarios.

* Update to test mocked http client

This needs a few eyes on it but the errors seen in tests appear to be due to a move in Psr7 from string bodies to one implementing Psr\Http\Message\StreamInterface so we need our mocked bodies to do the same.